aloe-vera: drop support
With the jump to a 5.15 kernel, we have removed support for the
MPFS-DEV-KIT. Feel free to use past releases of the SDK which do have
support for it, but going forward support for the HiFive Expansion Board
has been removed.

At least this means the muddied waters about the name "MPFS" have become
a little clearer!

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>