amp: update to the v2022.09 remoteproc configuration
remoteproc support has been implemented as part of the v2022.09 release.
Update to the latest revision of the amp examples to bring in this
support. A patch for buildroot is needed as the remoteproc firmware must
be deployed to the image, and thus the amp section in the Makefile is
dropped too.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>